Changes in parameters to make better images
In previous version, we had way too many quads, despite the fact that sources are galaxies. This was due to the fact that our sources were somewhat too small and faint. I also changed the way we are drawing the source-lens alignment. Finally, the source perturbations were too concentrated.

Summary of changes:
- larger and brighter sources
- source position drawn such that beta = R_s at maximum and maximum offset defined with respect to the Einstein radius
- source perturbations larger in size (less concentrated)